requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy>=2.0.25
alembic==1.12.1

# Authentication & Security
bcrypt==4.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# WebSocket & Real-time Communication
websockets==12.0
websocket-client==1.8.0

# Data Validation & Serialization
pydantic>=2.6.0

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Client (for testing)
requests==2.31.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dateutil==2.8.2 